Reduce usage of display: block
Many labels previously were `display: block`, so when you clicked
anywhere on the same line, you focused the associated input. This commit
fixes that by making the labels `display: inline` and wrapping them in a
div, so the inputs will only be focused when the actual inputs are
clicked.

The `.OrangeButton` style was also simplified.