Compile web workers with separate webpack project
Since `react-app-rewired` wasn't producing the correct production builds
(most of the files were not emitted), this commit uninstalls it, opting
to revert back to the original `react-scripts`.

To continue supporting web workers, this commit creates a new webpack
project under the `workers/` directory which compiles the workers and
emits them to `public/workers`. The worker compilation script is called
both before starting the dev server and building for production.

Note that unlike `src/`, making edits in `workers/` while the dev server
is running will not trigger a recompilation of the updated files. You
will have to manually run `npm run build-workers` to recompile the
updated files. Alternatively, you can restart the dev server, which will
also recompile the updated files.